Setup Auth0
This is the initial setup of Auth0. It's immensely ugly and in
need of refactoring, but if you click the Account button on the
top right, it will prompt you with an Auth0 login. I'll update
the README further to specify what setup needs to be done for local
development, but that will come in a separate commit. Just know it
works locally for me, so once I outline the steps it should work
for everybody else too.